Add files via upload
A library management system or automated library management system is a software that has been developed to handle the basic housekeeping functions in library. It helps to provide the information of books available to the staffs as well as to the customers. Library management system is designed in such a way that it can create, edit, delete and show the available books and users to the administrators and also to the users.